requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
apscheduler==3.11.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deprecated==1.2.15
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
orjson==3.10.12
packaging==24.2
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.2
pydantic-core==2.27.1
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
typer==0.13.1
typing-extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
wrapt==1.17.0
